Add mentor OTP verification and approval flow

+++ src/main/java/com/heartwise/backend/controller/MentorController.java
@@ -72,7 +72,7 @@ public class MentorController {
         map.put("specialty",      m.getSpecialty());
         map.put("skills",         m.getSkills());
         map.put("languages",      m.getLanguages());
-        map.put("price",          "₹" + (int) m.getPrice());
+        map.put("price",          "₹" + (double) m.getPrice());
         map.put("rating",         m.getRating());
         map.put("experience",     m.getExperience() + " yrs");
         map.put("gender",         m.getGender());
